pyproject.toml
[tool.poetry]
name = "nldb-app-test-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
pydantic = ">=2.7.4,<3.0.0"
langchain-openai = "^0.2.2"
ollama = "^0.3.3"
python-decouple = "^3.8"
langchain-core = "^0.3.10"
langchain-community = "^0.3.2"
langchain-ollama = "^0.2.0"
langserve = "^0.3.0"
pymysql = "^1.1.1"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
